Slider changed to not continuous
The slider was set to “continuous” which was causing the iPhone 4 to
stall for longer than a minute whenever a change was made to the
slider.  I set “continuous” to off so that the value for the slider is
passed when the user takes their finger off of the slider.  This
alleviated the stall.